+++ src/components/Questions/hooks/useQuestions.tsx
@@ -1,10 +1,24 @@
 import { useEffect } from "react";
+import { Form } from "antd";
 import { fetchQuestions } from "../../../redux/questions/api/asyncActions";
 import { moveToNextQuestion } from "../../../redux/questions/slice";
 import { useAppDispatch, useAppSelector } from "../../../redux/store";
 
 export const useQuestions = () => {
   const dispatch = useAppDispatch();
+
+  const [form] = Form.useForm();
+  const answer = Form.useWatch("answer", form);
+
+  const onSubmit = () => {
+    form.validateFields().then((data) => {
+      console.log("data", data);
+      nextQuestion();
+      form.resetFields();
+    });
+  };
+
+  const isDisabled = !answer?.length;
 
   useEffect(() => {
     dispatch(fetchQuestions());
@@ -21,6 +35,8 @@ export const useQuestions = () => {
   return {
     isLoading,
     currentQuestion,
-    nextQuestion,
+    form,
+    isDisabled,
+    onSubmit,
   };
 };
